feat: clean the code
