+++ src/components/button/style.ts
@@ -1,0 +1,18 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  button: {
+    paddingVertical: 12,
+    borderRadius: 37,
+    alignItems: 'center',
+    height: 79,
+    justifyContent: 'center',
+  },
+  buttonText: {
+    fontSize: 20,
+    fontWeight: '700',
+    fontFamily: 'MontserratRegular',
+  },
+});
+
+export default styles;
