+++ src/components/card/Card.tsx
@@ -1,8 +1,8 @@
-import { View, Text, StyleSheet } from 'react-native';
+import { View, Text } from 'react-native';
 import React from 'react';
-import COLOR from '../../constant/constant';
+import styles from './style';
 
-// Props ka Type Define Karein
+
 interface CardProps {
   title: string;
   subtitle: string;
@@ -20,34 +20,5 @@ const Card: React.FC<CardProps> = ({ title, subtitle, date, money }) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  card: {
-    backgroundColor: COLOR.quaternary,
-    padding: 20,
-    marginBottom: 10,
-    borderRadius: 20,
-    textAlign: 'justify',
-  },
-  title: {
-    fontSize: 29,
-    fontWeight: '800',
-    fontFamily: 'MontserratRegular',
-    color: COLOR.white,
-  },
-  datetext: {
-    fontSize: 14,
-    fontFamily: 'MontserratRegular',
-    fontWeight: '600',
-    color: COLOR.white,
-  },
-  moneytext: {
-    fontSize: 25,
-    fontWeight: '800',
-    fontFamily: 'MontserratRegular',
-    color: COLOR.primary,
-    marginTop: 5,
-  },
-});
 
 export default Card;
