+++ src/components/cardSection/style.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  scrollContainer: {
+    padding: 10,
+  },
+});
+
+export default styles;
