+++ src/components/customText/style.ts
@@ -1,0 +1,14 @@
+import {StyleSheet} from 'react-native';
+import COLOR from '../../constant/constant';
+
+const styles = StyleSheet.create({
+  text: {
+    fontSize: 24,
+    fontWeight: '700',
+    color: COLOR.primary,
+    textAlign: 'center',
+    fontFamily: 'MontserratRegular',
+  },
+});
+
+export default styles;
