+++ src/components/imgWithText/ImgWithText.tsx
@@ -1,35 +1,19 @@
-import {View, Text, Image, StyleSheet} from 'react-native';
 import React from 'react';
-import COLOR from '../../constant/constant';
+import {View, Text, Image, ImageSourcePropType} from 'react-native';
+import styles from './ImgWithTextStyles';
 
-const ImgWithText = ({imageSource, label}) => {
+interface ImgWithTextProps {
+  imageSource: ImageSourcePropType;
+  label: string;
+}
+
+const ImgWithText: React.FC<ImgWithTextProps> = ({imageSource, label}) => {
   return (
     <View style={styles.container}>
-      <View>
-        <Image source={imageSource} alt="profile" style={styles.image} />
-      </View>
+      <Image source={imageSource} style={styles.image} />
       <Text style={styles.title}>{label}</Text>
     </View>
   );
 };
 
-const styles = StyleSheet.create({
-  container: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    gap: 12,
-    height: 102,
-  },
-  image: {
-    width: 72,
-    height: 72,
-  },
-
-  title: {
-    fontSize: 14,
-    fontWeight: 600,
-    fontFamily: 'MontserratRegular',
-    color: COLOR.primary,
-  },
-});
 export default ImgWithText;
