+++ src/components/input/styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  input: {
+    width: '100%',
+    borderBottomWidth: 2,
+    borderBottomColor: '#101C1D',
+    paddingVertical: 8,
+    fontSize: 16,
+  },
+});
