+++ src/components/petCard/PetCard.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
-import {View, Text, StyleSheet, Image, TouchableOpacity} from 'react-native';
+import {View, Text, Image, TouchableOpacity} from 'react-native';
 import {PetCardProps} from '../../types/componentTypes';
+import {styles} from './styles';
+import MaterialIcons from 'react-native-vector-icons/MaterialIcons';
 import COLOR from '../../constant/constant';
 
 const PetCard: React.FC<PetCardProps> = ({
@@ -9,8 +11,7 @@ const PetCard: React.FC<PetCardProps> = ({
   age,
   location,
   gender,
-  icon, 
-  locationIcon,
+  icon,
   onPress,
   onIconPress,
 }) => {
@@ -29,16 +30,18 @@ const PetCard: React.FC<PetCardProps> = ({
             <Text style={styles.name}>{name}</Text>
             <Text style={styles.age}>{age}</Text>
             <View style={styles.locationRow}>
-              <Image source={locationIcon} style={styles.locationIcon} />
+              <MaterialIcons
+                name="location-on"
+                size={18}
+                color={COLOR.PRIMARY_RED}
+              />
               <Text style={styles.location}>{location}</Text>
             </View>
             <View style={styles.genderContainer}>
               <Text style={styles.gender}>{gender}</Text>
-              <View style={styles.genderContainer}>
-                <TouchableOpacity onPress={onIconPress}>
-                  <Image source={icon} style={styles.deleteIcon} />
-                </TouchableOpacity>
-              </View>
+              <TouchableOpacity onPress={onIconPress}>
+                <Image source={icon} style={styles.deleteIcon} />
+              </TouchableOpacity>
             </View>
           </View>
         </View>
@@ -46,106 +49,5 @@ const PetCard: React.FC<PetCardProps> = ({
     </TouchableOpacity>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    width: '100%',
-    height: 171,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#FFFFFF',
-    marginBottom: 12,
-    borderRadius: 10, 
-   
-},
-
-  imageContainer: {
-   width:200,
-    // height: ,
-    backgroundColor: '#C4C4C4',
-    borderRadius: 20,
-    overflow: 'hidden',
-    zIndex: 999,
-   
-  },
-  image: {
-    // width: '100%',
-    height: '100%',
-    resizeMode: 'cover',
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 4 }, 
-    shadowOpacity: 5,
-    shadowRadius: 6, 
-    elevation: 8, 
-  },
-  infoCard: {
-    flex: 1,
-    backgroundColor: '#FFFFFF',
-    borderRadius: 20, 
-    borderTopRightRadius:20,  
-    borderBottomRightRadius: 20,  
-    padding: 12,
-    marginLeft: -10,
-    height: 126,
-    width: 200,
-    shadowColor: '#000',
-    shadowOffset: {width: 0, height: 2},
-    shadowOpacity: 0.25,
-    shadowRadius: 3.84,
-    elevation: 5,
-  },
-  details: {
-    left: 10,
-  },
-  name: {
-    fontSize: 18,
-    fontWeight: '700',
-    color: COLOR.primary,
-    marginBottom: 2,
-  },
-  age: {
-    fontSize: 10,
-    color: COLOR.primary,
-    top: 2,
-    fontWeight: '500',
-  },
-  locationRow: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 8,
-    marginBottom: 4,
-    height: 30,
-  },
-  locationIcon: {
-    width: 9,
-    height: 13,
-    marginRight: 4,
-    tintColor: '#FF5A5F',
-  },
-  location: {
-    fontSize: 14,
-  },
-  genderContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-  },
-  gender: {
-    fontSize: 14,
-    color: '#666666',
-  },
-  deleteIcon: {
-    width: 24,
-    height: 24,
-    tintColor: 'red',
-  },
-  placeholder: {
-    width: 194,
-    height: 171,
-    backgroundColor: '#C4C4C4',
-    borderRadius: 20,
-  },
-});
 
 export default PetCard;
