+++ src/hooks/useProfile.ts
@@ -1,0 +1,18 @@
+import { useEffect } from "react";
+import { useSelector, useDispatch } from "react-redux";
+import { RootState, AppDispatch } from "../redux/store";
+import { fetchProfile } from "../redux/slices/profileImageSlice";
+
+const useProfile = () => {
+  const dispatch = useDispatch<AppDispatch>();
+  const profileData = useSelector((state: RootState) => state.profile.profileData);
+  const loading = useSelector((state: RootState) => state.profile.loading);
+
+  useEffect(() => {
+    dispatch(fetchProfile());
+  }, [dispatch]);
+
+  return { profileData, loading };
+};
+
+export default useProfile;
