+++ src/screens/adoptNowScreen/AdoptNowScreen.tsx
@@ -1,7 +1,7 @@
-import {View, Text,StyleSheet} from 'react-native';
+import {View, Text} from 'react-native';
 import React from 'react';
-import COLOR from '../../constant/constant';
-import ProfileCard from '../../components/profileCard/ProfileCard';
+import AdoptNowCard from '../../components/adoptNowCard/AdoptNowCard';
+import styles from './style';
 
 const AdoptNowScreen = () => {
   return (
@@ -9,25 +9,11 @@ const AdoptNowScreen = () => {
       <View>
         <Text style={styles.header}>Donation Request</Text>
       </View>
-      <View>
-      <ProfileCard />
+      <View style={styles.cardContainer}>
+        <AdoptNowCard />
       </View>
     </View>
   );
 };
 
-const styles = StyleSheet.create({
-    container: {
-      flex: 1,
-      backgroundColor:COLOR.white,
-      paddingHorizontal: 17,
-
-    },
-    header: {
-        fontSize: 24,
-        fontWeight:700,
-        marginTop: 30,
-        fontFamily: 'MontserratRegular',
-      },
-});
 export default AdoptNowScreen;
