+++ src/screens/donateScreen/DonateScreen.tsx
@@ -1,16 +1,10 @@
 import React from 'react';
-import {
-  View,
-  Text,
-  TouchableOpacity,
-  ScrollView,
-  StyleSheet,
-  Image,
-} from 'react-native';
-import CustomInput from '../../components/input/customInput';
+import {View, Text, TouchableOpacity, ScrollView, Image} from 'react-native';
+import CustomInput from '../../components/input/CustomInput';
 import PickerInput from '../../components/pickerInput/PickerInput';
-import COLOR from '../../constant/constant';
 import useDonateScreen from '../../hooks/useDonateScreen';
+import Icon from 'react-native-vector-icons/Ionicons';
+import styles from './style';
 
 const DonateScreen: React.FC<{navigation: any}> = ({navigation}) => {
   const {
@@ -44,7 +38,7 @@ const DonateScreen: React.FC<{navigation: any}> = ({navigation}) => {
       <TouchableOpacity
         style={styles.backButton}
         onPress={() => navigation.goBack()}>
-        <Text style={styles.backText}>{'←'}</Text>
+        <Icon name="arrow-back" size={24} color="#000" />
       </TouchableOpacity>
 
       <PickerInput
@@ -81,7 +75,7 @@ const DonateScreen: React.FC<{navigation: any}> = ({navigation}) => {
         selectedValue={vaccinated}
         onValueChange={setVaccinated}
         items={[
-          {label: '', value: ''},
+          {label: 'Select', value: ''},
           {label: 'Yes', value: 'yes'},
           {label: 'No', value: 'no'},
         ]}
@@ -131,7 +125,12 @@ const DonateScreen: React.FC<{navigation: any}> = ({navigation}) => {
       />
       <View>
         <TouchableOpacity style={styles.imageUpload} onPress={pickImage}>
-          {!imageUri && <Text style={styles.uploadText}>Select Image</Text>}
+          {!imageUri && (
+            <View style={styles.iconUpload}>
+              <Icon name="cloud-upload" size={24} color="#000" />
+              <Text style={styles.uploadText}>Select Image</Text>
+            </View>
+          )}
           {imageUri && (
             <Image source={{uri: imageUri}} style={styles.imagePreview} />
           )}
@@ -151,67 +150,5 @@ const DonateScreen: React.FC<{navigation: any}> = ({navigation}) => {
     </ScrollView>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 20,
-    backgroundColor: '#fff',
-  },
-  backButton: {
-    marginBottom: 30,
-  },
-  backText: {
-    fontSize: 46,
-    fontWeight: '900',
-    color: '#000',
-  },
-  label: {
-    fontSize: 18,
-    fontWeight: '600',
-    marginVertical: 5,
-    fontFamily: 'MontserratRegular',
-    color: COLOR.primary,
-    marginTop: 10,
-  },
-  imageUpload: {
-    borderWidth: 1,
-    borderStyle: 'dashed',
-    borderColor: '#000',
-    padding: 20,
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: '#F0F0F0',
-    height: 161,
-    borderRadius: 10,
-    marginTop: 20,
-  },
-  uploadText: {
-    fontSize: 16,
-    color: '#000',
-  },
-  imagePreview: {
-    width: '100%',
-    height: '100%',
-  },
-  button: {
-    backgroundColor: '#000',
-    padding: 15,
-    borderRadius: 10,
-    alignItems: 'center',
-    marginTop: 20,
-    marginBottom: 50,
-  },
-  buttonText: {
-    fontSize: 18,
-    color: '#fff',
-    fontWeight: 'bold',
-  },
-  errorText: {
-    color: 'red',
-    marginTop: 10,
-    textAlign: 'center',
-  },
-});
 
 export default DonateScreen;
