+++ src/screens/loginScreen/LoginScreen.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View, Text, TouchableOpacity, StyleSheet, KeyboardAvoidingView, ScrollView, Platform } from 'react-native';
 import { LoginScreenNavigationProp } from '../../types/navigation';
-import CustomInput from '../../components/input/customInput';
+import CustomInput from '../../components/input/CustomInput';
 import TermsCheckbox from '../../components/termCheckBox/TermCheckBox';
 import LoginButton from '../../components/button/CustomButton';
 import COLOR from '../../constant/constant';
